NEW CODE VERSION v2.0
This is an updated version of the code, following the recent changes in SciFor and DMFT_Tools packages especially
concerning the huge changes in the FFT and FFTGF modules for the Fourier transform (now FFTPACK).
This require to almost completely update the IPT library and the drivers.

Old drivers have been moved to a repo directory but kept for future maintenance.
Only the most important drivers for the Bethe lattice are kept in the drivers directory.
All the actual drivers have been tested. Yet possible bugs can emerge in the next future.
PAY ATTENTION!

remark: the real-axis superconducting algorithm use a formulation that is consistent with my
calculation but differs from the formulation in the ED code (where anyway it is clearly stated
that the two formulations are equivalent). I am keeping this now, waiting for further investigation.

todo: add a routine to measure observables in the MATSUBARA, REAL, KELDYSH formulation using the routines
already developed for the ED code. For the repulsive case it should be straightforward.
This new module will be necessary for different reasons: i) get more quantities to look at, ii) go to the multi-band
case, iii) [most important] use in the non-equilibrium DMFT-IPT code.