update: this version of the code fixes once and for all the solution using Keldysh formalism for both U>0 and U<0.
This method was bugged since the latest update concerning the
FFT method. The new FFT (based on FFTPACK) work fine, but
the use of 1:L arrays introduced a slight asymmetry in the real-time GF
which span -T:T-dt interval. This caused a miscalculation of products
of the form G(t)*G(-t) which ended up in an incorrect Self-energy.

A fix is introduced by allocating an L+1 array for the real-time GF
in 0:L. In order to use the same FFT an alternative version of the
real-time/real-freq. FFT are now introduced in IPT_GF. These handle
1:L (freq) and 0:L (time) arrays.

Solution was tested against IPT_SOC (real-axis) U>0 and former
Keldysh solution of the attractive.