Bug solved: there is an error in the FFT transforms implemented in DMFT_TOOLS_FFTGF, i.e. the module holding the FFT of the GF,Sigma using the scifor module SF_FFT_FFTPACK of plain FFT based on FFTPACK.
In order to solve this problem we took a new direction here.
We shall remove any reference to this module (DMFT_TOOLS_FFTGF) and keep
the FFT routines inside the code. This was already done for the real-frequency
real-time FFT for the Keldysh code (-L:L--> -T:T problem) and now it is
extended to the imaginary-freq. imaginary-time case.

So now the code entirely uses local FFT.
This kind of programming behavior has to be preferred with respect to having a
general module which in anycase can not stand all the possible cases presenting
in the different situations.

Thus we shall remove DMFT_TOOLS_FFTxxx (xxx=GF and AUX) from the DMFT_TOOLS pack.

Now the attractive code seems to work. Few other tests are in order anyway.