Fixed bug: The IPT for the attractive code was not working well. Benchmarks against the ED code has pointed us to a wrong sign in the FFT. We (Amaricci,Valli and Peronaci) discovered that there was a missing sign in the FFT of the anomalous components (when calling FFT with notail=T). This generated a complicated error path.
The results compare very well with ED now and are below the BCS results.
However we observe that a finite T the IPT solution does not recover exactly
the BCS results but remains near them, while converging to the 1st order Hartree
solution (i.e. the DMFT calculation retaining only the Hartree term).